+++ espnet2/bin/pack.py
@@ -1,7 +1,7 @@
 import argparse
 from typing import Type
 
-from espnet2.utils.pack_funcs import pack
+from espnet2.main_funcs.pack_funcs import pack
 
 
 class PackedContents:
